requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
attrs==22.1.0
    # via
    #   pytest
    #   pytest-mypy
black==22.12.0
    # via pytest-black
click==8.1.3
    # via
    #   -r ./requirements-pkg.txt
    #   black
coverage[toml]==6.5.0
    # via pytest-cov
exceptiongroup==1.0.4
    # via pytest
filelock==3.8.2
    # via pytest-mypy
flake8==4.0.1
    # via
    #   -r requirements.in
    #   pytest-flake8
iniconfig==1.1.1
    # via pytest
isort==5.11.2
    # via pytest-isort
mccabe==0.6.1
    # via flake8
mypy==0.991
    # via pytest-mypy
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==22.0
    # via pytest
pathspec==0.10.3
    # via black
platformdirs==2.6.0
    # via black
pluggy==1.0.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pytest==7.2.0
    # via
    #   -r requirements.in
    #   pytest-black
    #   pytest-cov
    #   pytest-flake8
    #   pytest-isort
    #   pytest-mypy
pytest-black==0.3.12
    # via -r requirements.in
pytest-cov==4.0.0
    # via -r requirements.in
pytest-flake8==1.1.1
    # via -r requirements.in
pytest-isort==3.1.0
    # via -r requirements.in
pytest-mypy==0.10.2
    # via -r requirements.in
pyyaml==6.0
    # via -r ./requirements-pkg.txt
toml==0.10.2
    # via pytest-black
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   mypy
    #   pytest
typing-extensions==4.4.0
    # via mypy
